* more modularization -> extract * refactor permission requests (TsFacade) to use a single request for all permissions instead of one per permission
