+++ bot/__init__.py
@@ -1,5 +1,6 @@
 from .TS3Bot import Bot
 from .audit_service import AuditService
 from .guild_service import GuildService
+from .user_service import UserService
 
-__all__ = ['AuditService', 'GuildService', 'Bot']
+__all__ = ['Bot', 'UserService', 'AuditService', 'GuildService']
