+++ bot/rest/controller/other_controller.py
@@ -3,7 +3,8 @@ import logging
 from flask import request
 from werkzeug.exceptions import abort
 
-from bot import Bot
+from bot import UserService
+from bot.commander_service import CommanderService
 from bot.rest.controller.abstract_controller import AbstractController
 from bot.rest.utils import try_get
 
@@ -11,14 +12,15 @@ LOG = logging.getLogger(__name__)
 
 
 class OtherController(AbstractController):
-    def __init__(self, bot: Bot):
+    def __init__(self, user_service: UserService, commander_service: CommanderService):
         super().__init__()
-        self._bot = bot
+        self._user_service = user_service
+        self._commander_service = commander_service
 
     def _routes(self):
         @self.api.route("/commanders", methods=["GET"])
         def _active_commanders():
-            acs = self._bot.getActiveCommanders()
+            acs = self._commander_service.get_active_commanders()
             return acs if acs is not None else abort(503)
 
         @self.api.route("/registration", methods=["DELETE"])
@@ -26,5 +28,5 @@ class OtherController(AbstractController):
             body = request.json
             gw2account = try_get(body, "gw2account", default="")
             LOG.info("Received request to delete user '%s' from the TS registration database.", gw2account)
-            changes = self._bot.removePermissionsByGW2Account(gw2account)
+            changes = self._user_service.delete_registration(gw2account)
             return {"changes": changes}
